refactor(era-validators): Using official lighthouse git repo
Instead of the Semiotic fork we are now using the sigp lighthouse repo. Needed to update values that are passed to certain functions, e.g. MerkleTree::create(). Also, the HistoricalSummary fields are private, so rather than passing in a HistoricalSummary for EthereumPostCapellaValidator, we are assuming that the block_summary_root is consumed directly.